requirements.txt
flask
flask-cors
pypdf2
langchain
langchain-community
langchain-groq
langchain-text-splitters
chromadb
gunicorn
groq
gspread
google-auth
